+++ packages/shared_preferences/shared_preferences_android/test/shared_preferences_android_test.dart
@@ -62,6 +62,12 @@ void main() {
       await testData.clear();
     });
 
+    test('registered instance', () {
+      SharedPreferencesAndroid.registerWith();
+      expect(SharedPreferencesStorePlatform.instance,
+          isA<SharedPreferencesAndroid>());
+    });
+
     test('getAll', () async {
       testData = InMemorySharedPreferencesStore.withData(kTestValues);
       expect(await store.getAll(), kTestValues);